Added ASL Smartsign Trigger
Added the Georgia Tech ASL Smartsign Trigger code to the renderer
module. Not sure how to test, so this might be a breaking feature, but
all the new code added is our trigger that works when by itself in a
userscript (aka it might just be a code placement issue).